Create new file collection and move class collection from file db to it Ulrica.bonnerstig@medieinstitutet.se Fredrik.skogberg@medieinstitutet.se Lukas.lennstring@medieinstitutet.se
